+++ src/global.cpp
@@ -10,11 +10,11 @@ Mesh::Mesh(const int mesh_precision)
 	BuildIcosaTable();
 }
 
-int Mesh::insert_vert(unsigned long long id1, unsigned long long id2, const Eigen::Vector3d& p)
+int Mesh::insert_vert(unsigned long long id1, unsigned long long id2, const Eigen::Vector3f& p)
 {
-	// vect3<double> tp(p);
+	// vect3<float> tp(p);
 	// vect3<long long> tmp = vect3f2vect3i(tp);
-	// // vect3<double> tmp(double(precise(p[0])), double(precise(p[1])), double(precise(p[2])));
+	// // vect3<float> tmp(float(precise(p[0])), float(precise(p[1])), float(precise(p[2])));
 	// if (vertices_map.find(tmp) == vertices_map.end())
 	// {
 	// 	verticesNum++;
@@ -64,14 +64,14 @@ void Mesh::reset()
 
 void Mesh::BuildIcosaTable()
 {
-	const double PI = 3.1415926f;
-	const double H_ANGLE = PI / 180 * 72;    // 72 degree = 360 / 5
-	const double V_ANGLE = atanf(1.0f / 2); 
-	double z, xy;                            // coords
-	double hAngle1 = -PI / 2 - H_ANGLE / 2;  // start from -126 deg at 1st row
-	double hAngle2 = -PI / 2;				// start from -90 deg at 2nd row
+	const float PI = 3.1415926f;
+	const float H_ANGLE = PI / 180 * 72;    // 72 degree = 360 / 5
+	const float V_ANGLE = atanf(1.0f / 2); 
+	float z, xy;                            // coords
+	float hAngle1 = -PI / 2 - H_ANGLE / 2;  // start from -126 deg at 1st row
+	float hAngle2 = -PI / 2;				// start from -90 deg at 2nd row
 
-	IcosaTable[0] = Eigen::Vector3d(0, 0, 1);
+	IcosaTable[0] = Eigen::Vector3f(0, 0, 1);
 	int i1, i2;
 
 	for (size_t i = 1; i <= 5; i++)
@@ -81,20 +81,20 @@ void Mesh::BuildIcosaTable()
 		z = std::sin(V_ANGLE);
 		xy = std::cos(V_ANGLE);
 
-		IcosaTable[i1] = Eigen::Vector3d(xy * cos(hAngle1), xy * sin(hAngle1), z);
-		IcosaTable[i2] = Eigen::Vector3d(xy * cos(hAngle2), xy * sin(hAngle2), -z);
+		IcosaTable[i1] = Eigen::Vector3f(xy * cos(hAngle1), xy * sin(hAngle1), z);
+		IcosaTable[i2] = Eigen::Vector3f(xy * cos(hAngle2), xy * sin(hAngle2), -z);
 
 		hAngle1 += H_ANGLE;
 		hAngle2 += H_ANGLE;
 	}
 
-	IcosaTable[11] = Eigen::Vector3d(0, 0, -1);	
+	IcosaTable[11] = Eigen::Vector3f(0, 0, -1);	
 }
 
-void Mesh::AppendSplash_ConstR(std::vector<Eigen::Vector3d>& splash_particles, const double radius)
+void Mesh::AppendSplash_ConstR(std::vector<Eigen::Vector3f>& splash_particles, const float radius)
 {
 	std::vector<int> tmp_vec_indices;
-	for (const Eigen::Vector3d& pos : splash_particles)
+	for (const Eigen::Vector3f& pos : splash_particles)
 	{
 		tmp_vec_indices.clear();
 		tmp_vec_indices.resize(12);
@@ -114,7 +114,7 @@ void Mesh::AppendSplash_ConstR(std::vector<Eigen::Vector3d>& splash_particles, c
 	}
 }
 
-void Mesh::AppendSplash_VarR(std::vector<Eigen::Vector3d>& splash_particles, std::vector<double>& splash_radius)
+void Mesh::AppendSplash_VarR(std::vector<Eigen::Vector3f>& splash_particles, std::vector<float>& splash_radius)
 {
 	std::vector<int> tmp_vec_indices;
 	for (int spi = 0; spi < splash_particles.size(); spi++)
